+++ Sources/SwiftUIKit/Extension/View+SafeArea.swift
@@ -19,8 +19,8 @@ public extension View {
      the previously ignored safe areas. Just use the binding
      to set the edge padding of the view you want to inset.
      */
-    func bindSafeAreaInsets(
-        for edge: Edge,
+    func bindSafeAreaInset(
+        of edge: Edge,
         to binding: Binding<CGFloat>) -> some View {
         self.bindGeometry(to: binding) {
             self.inset(for: $0, edge: edge)